pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "invoice-qc-service"
version = "1.0.0"
description = "Invoice Quality Control Service - Extract, validate, and quality check invoices"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "pdfplumber>=0.10.0",
    "PyPDF2>=3.0.0",
    "pydantic>=2.5.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "typer>=0.9.0",
    "python-dateutil>=2.8.2",
    "regex>=2023.10.3",
    "rich>=13.7.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
invoice-qc = "invoice_qc.cli.main:app"

[tool.setuptools]
packages = {find = {}}
package-dir = {"" = "."}

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

